chore: extract postgres-changes-filters suite into src/suites/postgres-changes-filters.ts
Part of REAL-1037.
